Added ApplicationSpecification and Transfer member in VulkanImage
- Removed few warnings in project

+++ VulkanCore/src/Platform/Vulkan/VulkanPipeline.cpp
@@ -31,7 +31,7 @@ namespace VulkanCore {
 			VkPushConstantRange pushConstantRange{};
 			pushConstantRange.stageFlags = VK_SHADER_STAGE_VERTEX_BIT | VK_SHADER_STAGE_FRAGMENT_BIT;
 			pushConstantRange.offset = 0;
-			pushConstantRange.size = pushConstantSize;
+			pushConstantRange.size = (uint32_t)pushConstantSize;
 
 			std::vector<VkDescriptorSetLayout> descriptorSetLayouts{ descriptorLayout.GetDescriptorSetLayout() };
